Update Questao 2.c

+++ Questao 2.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-float main(){
+int main(){
    
     printf("{-1,-2,-3,-4,-5}");
